requirements.txt
Flask==2.3.3
tensorflow==2.16.1
opencv-python-headless==4.10.0.84
numpy==1.24.3
Pillow==10.4.0
Werkzeug==3.0.1
gunicorn==21.2.0